day06 part 2

+++ day06/part1.scala
@@ -12,18 +12,18 @@ import scala.io.Source
     .reduce(_ * _)
   println(total)
 
-def distanceTravelledAt(time: Int, pushTime: Int) =
+def distanceTravelledAt(time: Long, pushTime: Long) =
   if time <= pushTime then 0
   else (time - pushTime) * (pushTime)
 
-case class Race(raceTime: Int, recordDistance: Int):
-  def winningSolutions: Int =
-    (0 to raceTime).count: pushTime =>
+case class Race(raceTime: Long, recordDistance: Long):
+  def winningSolutions: Long =
+    (0L to raceTime).count: pushTime =>
       distanceTravelledAt(time = raceTime, pushTime) > recordDistance
 
 def parseRaces(lines: Iterator[String]): Iterable[Race] =
-  val times = lines.next().split("""\s+""").drop(1).map(_.toInt)
-  val distances = lines.next().split("""\s+""").drop(1).map(_.toInt)
+  val times = lines.next().split("""\s+""").drop(1).map(_.toLong)
+  val distances = lines.next().split("""\s+""").drop(1).map(_.toLong)
   times
     .zip(distances)
     .map { (time, distance) => Race(time, distance) }
